Update dgtle.py
Multiple improvements:1. Since earlier this year, www.dgtle.com merged the web style from interest and article into the same format of url, and therefore the old way of extracting image location from webpage also needs to be changed.2. The server of Dgtle is very low quality, and the connection from North America to it seems very unstable. Shall there is a connection drop, I put a waiting period of 5 seconds.3. Sometimes, the poster uses an emoji from an external location, insert into the page as image, however, the url is completely different from the true image, and I have to remove those from the downloading list.4. For the same reason of 3, if the download url is separate apart from their location, then the download will have an error occurs. A new code added to remove those empty elements in the download list.5. Insead of changing the source code to change URL every time, an input function has been added, so users no longer need to modify the code, but only use the paste in the terminal to input the url.